refactor: simplify command policy helper
