+++ .agents/scripts/command_policy_evaluation.py
@@ -1,0 +1,212 @@
+#!/usr/bin/env python3
+# SPDX-License-Identifier: MIT
+# SPDX-FileCopyrightText: 2025-2026 Marcus Quinn
+"""Static and dynamic command-policy evaluation."""
+
+from __future__ import annotations
+
+import json
+import os
+import subprocess
+import sys
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any
+
+from command_policy_config import _decision
+from command_policy_matchers import _matches
+
+DECISION_RANK = {"allow": 0, "forbid": 1}
+
+
+@dataclass(frozen=True)
+class _EvaluationOptions:
+    guard_path: str = ""
+    worker: bool = False
+    worker_id: str = "unknown"
+    network_helper: str = ""
+
+
+def _evaluate_static(
+    invocations: list[list[str]], cwd: str, policy: dict[str, Any]
+) -> dict[str, Any]:
+    best = dict(policy["default_decision"])
+    best["schema_version"] = 2
+    for argv in invocations:
+        for rule in policy["rules"]:
+            if (
+                _matches(rule["matcher"], argv, cwd)
+                and DECISION_RANK[rule["decision"]] > DECISION_RANK[best["decision"]]
+            ):
+                best = _decision(rule["decision"], rule["id"], rule["reason"])
+    return best
+
+
+def _canonical_guard_path(
+    policy: dict[str, Any], explicit: str, script_dir: Path | None = None
+) -> Path:
+    if explicit:
+        return Path(explicit)
+    helper = next(
+        guard["helper"]
+        for guard in policy["dynamic_guards"]
+        if guard["kind"] == "canonical_git"
+    )
+    return (script_dir or Path(__file__).resolve().parent) / helper
+
+
+def _evaluate_canonical_git(
+    invocations: list[list[str]], cwd: str, guard: Path
+) -> dict[str, Any]:
+    git_invocations = [
+        argv for argv in invocations if argv and os.path.basename(argv[0]) == "git"
+    ]
+    if not git_invocations:
+        return _decision("allow", "git.no-invocation", "No Git invocation detected")
+    if not guard.is_file():
+        return _decision(
+            "forbid",
+            "git.guard-unavailable",
+            f"Canonical Git policy helper is unavailable: {guard}",
+        )
+    for argv in git_invocations:
+        try:
+            result = subprocess.run(
+                [
+                    sys.executable,
+                    str(guard),
+                    "--cwd",
+                    cwd,
+                    "--argv-json",
+                    json.dumps(argv[1:]),
+                ],
+                capture_output=True,
+                text=True,
+                timeout=10,
+                check=False,
+            )
+        except (OSError, subprocess.SubprocessError) as exc:
+            return _decision(
+                "forbid",
+                "git.guard-error",
+                f"Canonical Git policy check failed closed: {exc}",
+            )
+        if result.returncode != 0:
+            reason = result.stderr.strip() or (
+                f"Canonical Git guard failed with exit {result.returncode}"
+            )
+            return _decision("forbid", "git.canonical-worktree", reason)
+    return _decision(
+        "allow", "git.canonical-allow", "Canonical Git guard allowed every Git argv"
+    )
+
+
+def _network_guard_path(
+    policy: dict[str, Any], explicit: str, script_dir: Path | None = None
+) -> Path:
+    if explicit:
+        return Path(explicit)
+    override = os.environ.get("AIDEVOPS_NETWORK_TIER_HELPER", "")
+    if override:
+        return Path(override)
+    helper = next(
+        guard["helper"]
+        for guard in policy["dynamic_guards"]
+        if guard["kind"] == "worker_network"
+    )
+    return (script_dir or Path(__file__).resolve().parent) / helper
+
+
+def _evaluate_worker_network(
+    invocations: list[list[str]], cwd: str, helper: Path, worker_id: str
+) -> dict[str, Any]:
+    if not helper.is_file():
+        return _decision(
+            "forbid",
+            "network.helper-unavailable",
+            f"Required worker network policy helper is unavailable: {helper}",
+        )
+    for argv in invocations:
+        try:
+            result = subprocess.run(
+                [
+                    "/bin/bash",
+                    str(helper),
+                    "check-argv",
+                    json.dumps(argv),
+                    "--cwd",
+                    cwd,
+                    "--worker-id",
+                    worker_id,
+                ],
+                capture_output=True,
+                text=True,
+                timeout=10,
+                check=False,
+            )
+        except (OSError, subprocess.SubprocessError) as exc:
+            return _decision(
+                "forbid",
+                "network.helper-error",
+                f"Worker network policy failed closed: {exc}",
+            )
+        if result.returncode != 0:
+            reason = result.stderr.strip() or (
+                "Worker network policy denied or could not classify the command destination"
+            )
+            return _decision("forbid", "network.worker-policy", reason)
+    return _decision(
+        "allow", "network.worker-allow", "Worker network policy allowed every argv"
+    )
+
+
+def evaluate_invocations(
+    invocations: list[list[str]],
+    cwd: str,
+    policy: dict[str, Any],
+    *legacy_options: Any,
+    **named_options: Any,
+) -> dict[str, Any]:
+    options = _evaluation_options(legacy_options, named_options)
+    script_dir = Path(__file__).resolve().parent
+    decisions = [
+        _evaluate_static(invocations, cwd, policy),
+        _evaluate_canonical_git(
+            invocations,
+            cwd,
+            _canonical_guard_path(policy, options.guard_path, script_dir),
+        ),
+    ]
+    if options.worker:
+        decisions.append(
+            _evaluate_worker_network(
+                invocations,
+                cwd,
+                _network_guard_path(policy, options.network_helper, script_dir),
+                options.worker_id,
+            )
+        )
+    return max(decisions, key=lambda item: DECISION_RANK[item["decision"]])
+
+
+def _evaluation_options(
+    legacy_options: tuple[Any, ...], named_options: dict[str, Any]
+) -> _EvaluationOptions:
+    names = ("guard_path", "worker", "worker_id", "network_helper")
+    if len(legacy_options) > len(names):
+        raise TypeError(
+            f"evaluate_invocations() takes from 3 to 7 positional arguments "
+            f"but {len(legacy_options) + 3} were given"
+        )
+    values: dict[str, Any] = dict(zip(names, legacy_options))
+    for name, value in named_options.items():
+        if name not in names:
+            raise TypeError(
+                f"evaluate_invocations() got an unexpected keyword argument '{name}'"
+            )
+        if name in values:
+            raise TypeError(
+                f"evaluate_invocations() got multiple values for argument '{name}'"
+            )
+        values[name] = value
+    return _EvaluationOptions(**values)
